+++ app/Http/Resources/User.php
@@ -2,14 +2,33 @@
 
 namespace App\Http\Resources;
 
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 
 class User extends JsonResource
 {
     /**
+     * @var bool Indicates whether user permissions should be included or not.
+     */
+    private $withPermissions;
+
+    /**
+     * User resource constructor.
+     *
+     * @param \App\User $resource        The user model that should be transformed.
+     * @param bool      $withPermissions Indicates whether user permissions should be included or not (Default false).
+     */
+    public function __construct($resource, $withPermissions = false)
+    {
+        parent::__construct($resource);
+
+        $this->withPermissions = $withPermissions;
+    }
+
+    /**
      * Transform the resource into an array.
      *
-     * @param  \Illuminate\Http\Request $request
+     * @param  Request $request
      * @return array
      */
     public function toArray($request)
@@ -19,10 +38,12 @@ class User extends JsonResource
         }
 
         return [
-            'id'        => $this->id,
-            'firstname' => $this->firstname,
-            'lastname'  => $this->lastname,
-            'locale'    => $this->locale
+            'id'          => $this->id,
+            'firstname'   => $this->firstname,
+            'lastname'    => $this->lastname,
+            'locale'      => $this->locale,
+            'permissions' => $this->when($this->withPermissions, $this->permissions),
+            'modelName'   => $this->modelName
         ];
     }
 }