+++ database/seeds/DatabaseSeeder.php
@@ -12,5 +12,6 @@ class DatabaseSeeder extends Seeder
     public function run()
     {
         $this->call(RoomTypeSeeder::class);
+        $this->call(RolesAndPermissionsSeeder::class);
     }
 }